heaps code refactoring by removing testing methods out of the data structure.
