+++ src/main/java/org/clrs/algorithms/heaps/PriorityQueue.java
@@ -6,8 +6,6 @@ import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.StringJoiner;
 
-import org.clrs.algorithms.dc.Student;
-
 /**
  * An unbounded priority {@link Queue} based on a priority heap. The elements of
  * the priority queue are ordered according to their natural ordering, or by a
@@ -220,32 +218,4 @@ public class PriorityQueue<E> implements Queue<E> {
 		System.arraycopy(a, 1, arr, 0, size);
 		return arr;
 	}
-
-	public static void main(String[] args) {
-		final Queue<Integer> intQueue = PriorityQueue.of();
-		intQueue.insert(4);
-		intQueue.insert(2);
-		intQueue.insert(7);
-		intQueue.insert(5);
-
-		while (!intQueue.isEmpty())
-			System.out.println(intQueue.extract());
-
-		final Student bloch = new Student("Bloch", 3.81);
-		final Student jenny = new Student("Jenny", 2.51);
-		final Student meyers = new Student("Meyers", 3.76);
-		final Student stroustrup = new Student("Stroustrup", 3.94);
-		final Student forrest = new Student("Forrest", 2.74);
-
-		final Queue<Student> stdQueue = PriorityQueue.of(Comparator.comparingDouble(Student::getGpa).reversed());
-		stdQueue.insert(meyers);
-		stdQueue.insert(stroustrup);
-		stdQueue.insert(jenny);
-		stdQueue.insert(bloch);
-		stdQueue.insert(forrest);
-
-		while (!stdQueue.isEmpty())
-			System.out.println(stdQueue.extract());
-	}
-
 }
